Change string columns to text
I keep getting "Data too long" errors. I know that text performs worse
than string, esp. for temporary tables, e.g. when `ORDER BY` is used,
which for us is always, but I think that the impact will be marginal,
and the tradeoff is worth avoiding the need to debug length issues.